build: land g_spawn_mp
Adds src/game_mp/g_spawn_mp.cpp (913 lines) to the POSIX (and mirrored
Switch) targets.

Source touch-ups for clean compile under -Werror:
- g_spawn_mp.cpp:211, 648: KISAKHACK (uintptr_t) bridge casts on dvar
  integer-as-pointer for Scr_GetEntityAllowNull and g_motd

posix_backbone_stubs.cpp:
- 32 new function stubs covering G_NewString, G_SpawnItem,
  G_SpawnString, G_ParseSpawnVars, G_VehSpawner, G_VehCollmapSpawner,
  Scr_* family (GetType, AddFields, AddObject, FindField, GetObject,
  GetOffset, MakeArray, Set/GetAngles/Health/Origin, ExecThread/
  ExecEntThreadNum/AddExecThread, AddEntityNum/GetEntityRef/
  FreeEntityNum, SetStructField/SetDynamicEntityField,
  Get/SetHudElemField, FreeHudElemConstStrings, GetConstStringIncludeNull,
  NotifyNum), trigger_use, BG_FindItemForWeapon,
  Com_IsLegacyXModelName
- 3 dvar pointer globals: g_gravity, g_motd, plus g_hudelems[1024] storage
- remove 9 dup stubs (G_SpawnInt, G_SpawnFloat, G_LevelSpawnString,
  Scr_ExecEntThread, Scr_Notify, Scr_Get/SetGenericField,
  Scr_AddEntity, Scr_FreeEntity) now provided by g_spawn_mp.cpp

+++ src/game_mp/g_spawn_mp.cpp
@@ -208,7 +208,7 @@ void __cdecl Scr_SetGenericField(uint8_t *b, fieldtype_t type, int32_t ofs)
         *(float *)&b[ofs + 8] = vec[2];
         break;
     case F_ENTITY:
-        *(uint32_t *)&b[ofs] = (uint32_t)Scr_GetEntityAllowNull(0);
+        *(uint32_t *)&b[ofs] = (uint32_t)(uintptr_t)Scr_GetEntityAllowNull(0);
         break;
     case F_ENTHANDLE:
         pEnt = (EntHandle *)&b[ofs];
@@ -645,7 +645,7 @@ void __cdecl SP_worldspawn()
     }
     G_LevelSpawnString("message", "", &s);
     SV_SetConfigstring(3, (char *)s);
-    SV_SetConfigstring(10, (char *)g_motd->current.integer);
+    SV_SetConfigstring(10, (char *)(uintptr_t)g_motd->current.integer);
     G_LevelSpawnString("gravity", "800", &s);
     if (!g_gravity)
         MyAssertHandler(".\\game_mp\\g_spawn_mp.cpp", 1123, 0, "%s", "g_gravity");
